+++ src/components/common/OneDayList/style.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import { S_FlexBox } from '../../../style/FlexBox.style';
+
+export const S_OneDayListtWrapper = styled(S_FlexBox)`
+  width: 90%;
+  justify-content: space-evenly;
+  flex-wrap: wrap;
+`;
